+++ dinky-web/src/pages/DevOps/JobDetail/JobMetrics/JobChart/JobChart.tsx
@@ -23,7 +23,7 @@ import { SseData } from '@/models/Sse';
 import { SSE_TOPIC } from '@/pages/DevOps/constants';
 import { JobMetricsItem, MetricsTimeFilter } from '@/pages/DevOps/JobDetail/data';
 import { getMetricsData } from '@/pages/DevOps/JobDetail/srvice';
-import { ChartData } from '@/pages/Metrics/Job/data';
+import { ChartData } from '@/pages/Metrics/JobMetricsList/data';
 import { MetricsDataType } from '@/pages/Metrics/Server/data';
 import { Jobs } from '@/types/DevOps/data';
 import { Empty, Row, Spin } from 'antd';
